--- src/hero/Hero.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import Heading from './heading/Heading';
-import Preview from './preview/Preview';
-
-const Hero = () => (
-  <>
-    <Heading />
-    <Preview />
-  </>
-);
-export default Hero;
