--- src/hero/preview/Preview.style.js
@@ -1,22 +1,0 @@
-import styled from 'styled-components';
-
-export const Section = styled.section`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 46px 0 0 0;
-`;
-
-export const SubReddit = styled.p`
-  font-weight: 500;
-  line-height: 19.5px;
-  margin: 0;
-  padding: 0 2px 1px 0;
-  text-align: center;
-`;
-
-export const Table = styled.img`
-  height 332px;
-  margin: 36px auto;
-  width: 1114px;
-`;
