+++ src/page-home/HomePage.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Hero from './Hero';
+import Info from './Info';
+
+const HomePage = () => (
+  <main>
+    <Hero />
+    <Info />
+  </main>
+);
+
+export default HomePage;
